implement the table filter functionality.
