+++ ambillion-ui/src/utils/table/columns.tsx
@@ -1,0 +1,511 @@
+import { TableColumn } from 'react-data-table-component';
+
+interface DataRow {
+    id: number;
+    firstName: string;
+    lastName: string;
+    companyName: string;
+    mobileNumber: string;
+    emailAddress: string;
+    isVerified: boolean;
+}
+export const userTableColumns: TableColumn<DataRow>[] = [
+    {
+        name: 'ID',
+        selector: (row) => row.id,
+        sortable: true,
+        width: '60px'
+    },
+    {
+        name: 'FirstName',
+        selector: (row) => row.firstName,
+        sortable: true
+    },
+    {
+        name: 'LastName',
+        selector: (row) => row.lastName,
+        sortable: true
+    },
+    {
+        name: 'CompanyName',
+        selector: (row) => row.companyName,
+        sortable: true
+    },
+    {
+        name: 'MobileNumber',
+        selector: (row) => row.mobileNumber,
+        sortable: true
+    },
+    {
+        name: 'EmailAddress',
+        selector: (row) => row.emailAddress,
+        sortable: true
+    },
+    {
+        name: 'Verified',
+        selector: (row) => row.isVerified,
+        sortable: true,
+        cell: (row) =>
+            row.isVerified ? (
+                <span className="badge bg-success-subtle text-success rounded fw-semibold p-2">
+                    Verified
+                </span>
+            ) : (
+                <span className="badge bg-danger-subtle text-danger rounded fw-semibold p-2">
+                    UnVerified
+                </span>
+            )
+    }
+];
+export const dummyUserTableData = [
+    {
+        id: 1,
+        firstName: 'Vishal',
+        lastName: 'Devani',
+        companyName: 'ABC Inc.',
+        mobileNumber: '1234567890',
+        emailAddress: 'vishal.devani@example.com',
+        isVerified: true
+    },
+    {
+        id: 2,
+        firstName: 'John',
+        lastName: 'Smith',
+        companyName: 'XYZ Corp.',
+        mobileNumber: '2345678901',
+        emailAddress: 'john.smith@example.com',
+        isVerified: false
+    },
+    {
+        id: 3,
+        firstName: 'Emily',
+        lastName: 'Brown',
+        companyName: '123 Company',
+        mobileNumber: '3456789012',
+        emailAddress: 'emily.brown@example.com',
+        isVerified: true
+    },
+    {
+        id: 4,
+        firstName: 'Michael',
+        lastName: 'Johnson',
+        companyName: '456 Corp.',
+        mobileNumber: '4567890123',
+        emailAddress: 'michael.johnson@example.com',
+        isVerified: true
+    },
+    {
+        id: 5,
+        firstName: 'Sarah',
+        lastName: 'Wilson',
+        companyName: 'DEF Ltd.',
+        mobileNumber: '5678901234',
+        emailAddress: 'sarah.wilson@example.com',
+        isVerified: false
+    },
+    {
+        id: 6,
+        firstName: 'David',
+        lastName: 'Lee',
+        companyName: '789 Corp.',
+        mobileNumber: '6789012345',
+        emailAddress: 'david.lee@example.com',
+        isVerified: false
+    },
+    {
+        id: 7,
+        firstName: 'Emma',
+        lastName: 'Garcia',
+        companyName: 'GHI Inc.',
+        mobileNumber: '7890123456',
+        emailAddress: 'emma.garcia@example.com',
+        isVerified: true
+    },
+    {
+        id: 8,
+        firstName: 'Daniel',
+        lastName: 'Martinez',
+        companyName: 'JKL Corp.',
+        mobileNumber: '8901234567',
+        emailAddress: 'daniel.martinez@example.com',
+        isVerified: true
+    },
+    {
+        id: 9,
+        firstName: 'Sophia',
+        lastName: 'Lopez',
+        companyName: 'MNO Ltd.',
+        mobileNumber: '9012345678',
+        emailAddress: 'sophia.lopez@example.com',
+        isVerified: false
+    },
+    {
+        id: 10,
+        firstName: 'James',
+        lastName: 'Davis',
+        companyName: 'PQR Inc.',
+        mobileNumber: '0123456789',
+        emailAddress: 'james.davis@example.com',
+        isVerified: false
+    },
+    {
+        id: 11,
+        firstName: 'Olivia',
+        lastName: 'Miller',
+        companyName: 'STU Corp.',
+        mobileNumber: '9876543210',
+        emailAddress: 'olivia.miller@example.com',
+        isVerified: true
+    },
+    {
+        id: 12,
+        firstName: 'William',
+        lastName: 'Rodriguez',
+        companyName: 'VWX Ltd.',
+        mobileNumber: '8765432109',
+        emailAddress: 'william.rodriguez@example.com',
+        isVerified: true
+    },
+    {
+        id: 13,
+        firstName: 'Ava',
+        lastName: 'Martinez',
+        companyName: 'YZA Inc.',
+        mobileNumber: '7654321098',
+        emailAddress: 'ava.martinez@example.com',
+        isVerified: false
+    },
+    {
+        id: 14,
+        firstName: 'Alexander',
+        lastName: 'Hernandez',
+        companyName: 'BCD Corp.',
+        mobileNumber: '6543210987',
+        emailAddress: 'alexander.hernandez@example.com',
+        isVerified: true
+    },
+    {
+        id: 15,
+        firstName: 'Mia',
+        lastName: 'Gonzalez',
+        companyName: 'EFG Ltd.',
+        mobileNumber: '5432109876',
+        emailAddress: 'mia.gonzalez@example.com',
+        isVerified: true
+    },
+    {
+        id: 16,
+        firstName: 'Ethan',
+        lastName: 'Perez',
+        companyName: 'HIJ Inc.',
+        mobileNumber: '4321098765',
+        emailAddress: 'ethan.perez@example.com',
+        isVerified: true
+    },
+    {
+        id: 17,
+        firstName: 'Isabella',
+        lastName: 'Sanchez',
+        companyName: 'KLM Corp.',
+        mobileNumber: '3210987654',
+        emailAddress: 'isabella.sanchez@example.com',
+        isVerified: false
+    },
+    {
+        id: 18,
+        firstName: 'Alexander',
+        lastName: 'Gonzalez',
+        companyName: 'NOP Ltd.',
+        mobileNumber: '2109876543',
+        emailAddress: 'alexander.gonzalez@example.com',
+        isVerified: false
+    },
+    {
+        id: 19,
+        firstName: 'Sophia',
+        lastName: 'Rivera',
+        companyName: 'QRS Inc.',
+        mobileNumber: '1098765432',
+        emailAddress: 'sophia.rivera@example.com',
+        isVerified: false
+    },
+    {
+        id: 20,
+        firstName: 'Mason',
+        lastName: 'Brooks',
+        companyName: 'TUV Corp.',
+        mobileNumber: '0987654321',
+        emailAddress: 'mason.brooks@example.com',
+        isVerified: true
+    },
+    {
+        id: 21,
+        firstName: 'Amelia',
+        lastName: 'Roberts',
+        companyName: 'WXY Ltd.',
+        mobileNumber: '9876543210',
+        emailAddress: 'amelia.roberts@example.com',
+        isVerified: true
+    },
+    {
+        id: 22,
+        firstName: 'Logan',
+        lastName: 'Cook',
+        companyName: 'ZAB Inc.',
+        mobileNumber: '8765432109',
+        emailAddress: 'logan.cook@example.com',
+        isVerified: false
+    },
+    {
+        id: 23,
+        firstName: 'Evelyn',
+        lastName: 'Bailey',
+        companyName: 'CDE Corp.',
+        mobileNumber: '7654321098',
+        emailAddress: 'evelyn.bailey@example.com',
+        isVerified: false
+    },
+    {
+        id: 24,
+        firstName: 'Lucas',
+        lastName: 'James',
+        companyName: 'FGH Ltd.',
+        mobileNumber: '6543210987',
+        emailAddress: 'lucas.james@example.com',
+        isVerified: true
+    },
+    {
+        id: 25,
+        firstName: 'Abigail',
+        lastName: 'Phillips',
+        companyName: 'IJK Inc.',
+        mobileNumber: '5432109876',
+        emailAddress: 'abigail.phillips@example.com',
+        isVerified: false
+    },
+    {
+        id: 26,
+        firstName: 'Jackson',
+        lastName: 'Evans',
+        companyName: 'LMN Corp.',
+        mobileNumber: '4321098765',
+        emailAddress: 'jackson.evans@example.com',
+        isVerified: false
+    },
+    {
+        id: 27,
+        firstName: 'Charlotte',
+        lastName: 'Howard',
+        companyName: 'OPQ Ltd.',
+        mobileNumber: '3210987654',
+        emailAddress: 'charlotte.howard@example.com',
+        isVerified: true
+    },
+    {
+        id: 28,
+        firstName: 'Daniel',
+        lastName: 'Morales',
+        companyName: 'RST Inc.',
+        mobileNumber: '2109876543',
+        emailAddress: 'daniel.morales@example.com',
+        isVerified: true
+    },
+    {
+        id: 29,
+        firstName: 'Harper',
+        lastName: 'Ward',
+        companyName: 'UVW Corp.',
+        mobileNumber: '1098765432',
+        emailAddress: 'harper.ward@example.com',
+        isVerified: true
+    },
+    {
+        id: 30,
+        firstName: 'Aiden',
+        lastName: 'Torres',
+        companyName: 'XYZ Ltd.',
+        mobileNumber: '0987654321',
+        emailAddress: 'aiden.torres@example.com',
+        isVerified: true
+    },
+    {
+        id: 31,
+        firstName: 'Madison',
+        lastName: 'Flores',
+        companyName: '123 Inc.',
+        mobileNumber: '9876543210',
+        emailAddress: 'madison.flores@example.com',
+        isVerified: true
+    },
+    {
+        id: 32,
+        firstName: 'Elijah',
+        lastName: 'Nguyen',
+        companyName: '456 Corp.',
+        mobileNumber: '8765432109',
+        emailAddress: 'elijah.nguyen@example.com',
+        isVerified: true
+    },
+    {
+        id: 33,
+        firstName: 'Avery',
+        lastName: 'King',
+        companyName: '789 Ltd.',
+        mobileNumber: '7654321098',
+        emailAddress: 'avery.king@example.com',
+        isVerified: true
+    },
+    {
+        id: 34,
+        firstName: 'Grayson',
+        lastName: 'Scott',
+        companyName: 'ABC Inc.',
+        mobileNumber: '6543210987',
+        emailAddress: 'grayson.scott@example.com',
+        isVerified: true
+    },
+    {
+        id: 35,
+        firstName: 'Scarlett',
+        lastName: 'Gomez',
+        companyName: 'XYZ Corp.',
+        mobileNumber: '5432109876',
+        emailAddress: 'scarlett.gomez@example.com',
+        isVerified: true
+    },
+    {
+        id: 36,
+        firstName: 'Logan',
+        lastName: 'Hill',
+        companyName: '123 Company',
+        mobileNumber: '4321098765',
+        emailAddress: 'logan.hill@example.com',
+        isVerified: true
+    },
+    {
+        id: 37,
+        firstName: 'Lily',
+        lastName: 'Perez',
+        companyName: '456 Corp.',
+        mobileNumber: '3210987654',
+        emailAddress: 'lily.perez@example.com',
+        isVerified: true
+    },
+    {
+        id: 38,
+        firstName: 'Owen',
+        lastName: 'Collins',
+        companyName: '789 Ltd.',
+        mobileNumber: '2109876543',
+        emailAddress: 'owen.collins@example.com',
+        isVerified: true
+    },
+    {
+        id: 39,
+        firstName: 'Liam',
+        lastName: 'Stewart',
+        companyName: 'ABC Inc.',
+        mobileNumber: '1098765432',
+        emailAddress: 'liam.stewart@example.com',
+        isVerified: true
+    },
+    {
+        id: 40,
+        firstName: 'Zoe',
+        lastName: 'Mitchell',
+        companyName: 'XYZ Corp.',
+        mobileNumber: '0987654321',
+        emailAddress: 'zoe.mitchell@example.com',
+        isVerified: true
+    },
+    {
+        id: 41,
+        firstName: 'Jackson',
+        lastName: 'Rogers',
+        companyName: '123 Company',
+        mobileNumber: '9876543210',
+        emailAddress: 'jackson.rogers@example.com',
+        isVerified: true
+    },
+    {
+        id: 42,
+        firstName: 'Chloe',
+        lastName: 'Price',
+        companyName: '456 Corp.',
+        mobileNumber: '8765432109',
+        emailAddress: 'chloe.price@example.com',
+        isVerified: true
+    },
+    {
+        id: 43,
+        firstName: 'Carter',
+        lastName: 'Ward',
+        companyName: '789 Ltd.',
+        mobileNumber: '7654321098',
+        emailAddress: 'carter.ward@example.com',
+        isVerified: true
+    },
+    {
+        id: 44,
+        firstName: 'Emma',
+        lastName: 'Brooks',
+        companyName: 'ABC Inc.',
+        mobileNumber: '6543210987',
+        emailAddress: 'emma.brooks@example.com',
+        isVerified: true
+    },
+    {
+        id: 45,
+        firstName: 'Aiden',
+        lastName: 'Turner',
+        companyName: 'XYZ Corp.',
+        mobileNumber: '5432109876',
+        emailAddress: 'aiden.turner@example.com',
+        isVerified: true
+    },
+    {
+        id: 46,
+        firstName: 'Grace',
+        lastName: 'Scott',
+        companyName: '123 Company',
+        mobileNumber: '4321098765',
+        emailAddress: 'grace.scott@example.com',
+        isVerified: true
+    },
+    {
+        id: 47,
+        firstName: 'Ethan',
+        lastName: 'Gonzalez',
+        companyName: '456 Corp.',
+        mobileNumber: '3210987654',
+        emailAddress: 'ethan.gonzalez@example.com',
+        isVerified: true
+    },
+    {
+        id: 48,
+        firstName: 'Hannah',
+        lastName: 'Evans',
+        companyName: '789 Ltd.',
+        mobileNumber: '2109876543',
+        emailAddress: 'hannah.evans@example.com',
+        isVerified: true
+    },
+    {
+        id: 49,
+        firstName: 'Luke',
+        lastName: 'Ward',
+        companyName: 'ABC Inc.',
+        mobileNumber: '1098765432',
+        emailAddress: 'luke.ward@example.com',
+        isVerified: true
+    },
+    {
+        id: 50,
+        firstName: 'Leah',
+        lastName: 'Morris',
+        companyName: 'XYZ Corp.',
+        mobileNumber: '0987654321',
+        emailAddress: 'leah.morris@example.com',
+        isVerified: true
+    }
+];
